+++ lib/languages/language.rb
@@ -56,7 +56,7 @@ module Languages
     end
 
     def <=>(other)
-      other.iso639_3 <=> iso639_3
+      iso639_3 <=> other.iso639_3
     end
   end
 end