+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Add `csv` gem as dependency as it is no longer part of stdlib with Ruby 3.4.
+
 ## [0.10.0] - 2025-08-12
 
 ### Removed